Small features and modifications:
1. Import eagle parts: Create new parts and packages if it doesn't exist.
All you need to do is to (later) fill in the informations (height of
parts, feeder location,..)

2. Bestfit board location.
Use four points on the board, it tries to find the best rotation angle and
offset to get the lowest total error.

3. Sorting pick order.
Sort the pick order according to the jobboard table. I also added "feeder"
column and "progress" column. Sorting based on feeder is useful so you can
"take-care" of cut-tapes one at a time. Todo (later): stop and ask before
continuing with new feeder.

4. Some small buttons:
- Position camera/tool to the selected part in job-panel. This is useful
  for bestfit board-positionning. Just select the part, click to go there,
  adjust the position, and click next.
  - Feed, pick and place buttons.

5. Driver for linux-cnc.
  Setup ip address/port in machine.xml