OpenBuildsDriver, GcodeDriver and LinuxCNC disconnect when disabled.
This allows for a serial port to be reused in another program (i.e. Pronterface) or unplugged, or the controller reset without closing OpenPNP. If the serial port becomes available again, the machine can be enabled again and the connection succeeds.

Before this change this failed at least under Windows where a COM port name is blocked until explicitly closed by the process. Even resetting/re-plugging the controller will not free the port and also not reestablished the connection. OpenPNP has to be closed and the controller reset/re-plugged one more time, before a freshly started OpenPNP can reconnect.

When closing OpenPNP the user loses the GUI navigation which is very annoying when editing Gcode fragments. Along with Gcode editing, users may need to change config.txt on Smoothieware controllers, which requires a reset of the controller. Which again (at least on Windows) was only possible after exiting OpenPNP.