Update JobPlacementsPanel.java
Change the functionality of the Move camera to next placement button.
Old behavior: The camera would move to the current location and the highlighted part would move to the next part.
New behavior: The highlighted part is moved and then the camera is moved to its location.
This way, the part in the camera view and the part highlighted in the parts list are the same. Then if you adjust the placement, when you press the "Set placement to current camera location" etc the correct part will be edited instead of the next part in the list.