In order to recalibrate nozzle tips on homing, we need to get the home() method call to the nozzle tips.
This required adding a home() method to NozzleTip and the home() recursion to AbstractMachine.

The first attempt was adding a machineHeadActivity() machineEnabled() listener. But these are not Exception enabled and we need to be able to throw Exceptions and abort the homing, if calibration is configured as a prerequisite of homing and then fails.

The original PR#825 already added home() recursion into all the HeadMountables for the same reason. This is now actually unused, as the homing has moved from the nozzles to the now global nozzle tips as they are directly attached to the machine. I still prefer keeping the HeadMountable.home() recursion, as a stepping stone to move visual homing from the GCodeDriver (where it doesn't belong) to the down-looking Camera.