Renamed "Z Probe" section of ReferenceHeadConfigurationWizard to "Z Actuators".
Renamed existing "Z Probe Actuator Name" to "Z Contact Actuator Name" to
distinguish it from the added "Safe Z Actuator Name".  "Safe Z" is an
actuator that can be used to confirm that the nozzle is sucessfully
retracted.  "Z Contact" is being worked on in PR #855 and identifies
when the z-axis has hit something with the tip of the nozzle.

Added a check in AbstractHead to check the "Safe Z" actuator status in
moveToSafeZ, though this should be moved to the nozzle.