ReferenceDragFeeder: Rename actuator to dragPinActuator to distinguish it from the other actuators in the class.
For CHMT it's better to try to retract the drag pin before the backoff
move is made.  This gives the spring some time to do its job.

Adds a safety check to confirm the drag pin was sucessfully retracted.
If not it tries to a second backoff move by the same amount to try to
free the pin.  I don't think this is bad because the drag pin is
obviously already in a binding position.

Removed unneeded moveToSafeZ calls.


ReferenceDragFeederConfigurationWizard:
Added UI elements to allow user to identify actuator that reads the
status of the drag pin's retraction and what that value should be
compared to.