Adds support for vacuum / blow level
If set on the part's package vacuum level and / or blow level
will be used in pick and place operations instead of just using
on / off.

Changes Package and Nozzle, in addition to adding support in
the neoden4 driver:

The neoden4 has individual (per nozzle) support for vacuum / blow
level, this is meant to be used per part to unsure proper pickup
and proper place (some parts needs a bit of air to be released
from the nozzle).