Starting the implementation of a reusable CvPipeline configuration system as suggested in #843, so WIP
Creating a new class Pipeline
-> includes a CvPipeline
-> includes a Camera
-> is saved to XML with an id and referenced by that

The actual implementations are in ReferenceBottomPipeline and ReferenceStripFeederPipeline as they load a specific default implementation. (ReferenceBottomPipeline is not used but shows the concept of having a bottom down camera)

This solves assigning the same pipeline configuration to multiple feeders. For now this is located in the ReferenceStripFeeder but should be moved up in the hierarchy

This basically also solves the issue with having multiple up cameras since a pipeline is now tied to a camera and thus could be used in job processing

This makes sense to me since the pipeline settings are camera dependent