Add read for http actuator (#1130)
* HTTP Actuator Read URL addition

* Fix dialog order

* Add regexp handling for HTTP Actuator

* Add regex handling for HttpActuator read

The regex is empty per default and the output is just passed through.
If the regex is matched it's the same as in the gdcode driver that
the matched output is produced. This is usefull if the returned
page contains several return values.

* Add a unit test to HttpActuator read

The unit test provides a sample web server (built in from java)
and provides one url with a value and the value is checked that
it matched the regex.