Allow for pick angle tolerance under limited articulation.
Some feeders use vision to determine the precise pick location and
angle. Under limited articulation Rotation Mode we need to predict not
only the pick-to-place articulation but also the needed leeway on both
the pick and the alignment/placement side. This adds it on the pick side
(it was previously only present omn the alignment side).