Abandon simple cache in favor of a KeyedObjectPool
This commit introduces a pooled approach to script engine caching to avoid issues that could occur if only a single scripting engine was used for concurrent / nested invocations.
With the pooled approach, engines that are borrowed cannot be re-used until they are given back to the pool, leading to creation of another engine in the pool if two scripts need to run at the same time with the same engine type.