PhotonFeeder: Add vision support for 0402 components.
The problem with 0402 components is that the tape is moved forward by 2mm while
the hole pitch is of 4mm. Thus we have to account for changes in the linear part
of the hole-to-part offset, which needs to also be forwarded and then truncated
to the hole pitch size.

This code will update the hole-to-part linear offset each time a new component
is fed, by default it would start at 2mm. The hole-yo-part offset is reset each
time the hole to part offset is updated.