Remove overhead by using ReferenceDriverCommunications::writeBytes()
Previously, write() was called in a loop over all bytes.
This seems unnecessary because that method will in turn
call writeBytes() with a length-1 array for each byte.

Now we're exporting writeBytes() as a public method so
the machine drivers can use it directly.

Neoden4Driver was adapted to use writeBytes(). Without
this change, serial communication would be to slow on
the original hardware and the machine controller would
not recognize command sequences.