For corner shots, add whole part mask to the pipeline and parametrize it from bottom vision.
In Vision Composition a.k.a. multi-shot bottom vision, the corner shots
are done with the nozzle holding the part off-center of the camera. This
means that another nozzle (with part) might come into the view of the
camera and disrupt the detection. Similarly, some machines might have a
limited size backdrop shade, so machine parts with bright reflections
might enter the view. This PR adds a second "part mask" to the pipeline
so the camera view is typically half masked off.