gcodeasyncdriver: change condition variable for confirmationComplete to the formally correct format
The previously implemented condition variable is very confusing for no benefit.
The formally correct way is to lock the loop preheader and header while unlocking in the body.
Change to the formally correct so I don't have to guess if the volatile usage is correct or not.

This also fixes a data race on errorResponse (and arguably confirmationComplete but it's volatile and not everyone would agree).