photonfeeder: while doing moveWhileFeeding move to safeZ rather than 0
This fixes an interesting case where my nozzles go above safeZ
while doing a moveWhileFeeding action.

At first I wanted to move a couple mms above the pick location but
this is has a lot of tiny details to get right along of being bugged,
so this makes more sense.